Use `uv` to manage python operator venvs
This change replaces the use of the `venv` and `pip` modules with
`uv`. The latter is a lot faster, especially for creating new
venvs, because it does not bootstrap pip implicitly.

With my local measurements I get around 0.8s - 1.0s creation overhead
when not requiring additional modules. This is a large improvement
compared to `python -m venv && python -m pip install ...`, which
often takes more than 10 seconds.

Thanks to that we are able to remove the venv caching logic and
make all venvs ephemeral from now on. This means that repeated
startups of the operator are now _slower_, because a cache hit
avoided venv creation and dependency intallation entirely.

We still decided to remove the caching due to the recurring issues
with the intricate interprocess synchronization logic.

An additional benefit of this is that restarting a python operator
now automatically updates the dependencies to their latest versions.